// apps/backend/notification-service/src/notification/notification.type.ts
export enum NotificationChannel {
	EMAIL = 'email',
	SMS = 'sms',
	SYSTEM = 'system',
}

export enum NotificationType {
	ALREADY_REGISTERED = 'already-registered',
	VERIFICATION = 'verification',
	WELCOME = 'welcome',
	PASSWORD_RESET = 'password-reset',
	GENERIC = 'generic',
	INVITE_TEAM = 'invite-team',
}